+++ eslint.config.mjs
@@ -1,12 +1,19 @@
 import eslint from '@eslint/js'
 import tseslint from 'typescript-eslint'
 import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended'
+import { createTypeScriptImportResolver } from 'eslint-import-resolver-typescript'
+import { createNodeResolver, importX } from 'eslint-plugin-import-x'
+import { globalIgnores } from 'eslint/config'
 
 export default tseslint.config(
   eslint.configs.recommended,
   tseslint.configs.recommended,
   eslintPluginPrettierRecommended,
+  importX.flatConfigs.recommended,
   {
+    env: {
+      browser: true
+    },
     rules: {
       'no-unused-vars': 'off',
       '@typescript-eslint/consistent-type-imports': 'error',
@@ -25,7 +32,11 @@ export default tseslint.config(
       '@typescript-eslint/no-explicit-any': 'off',
       '@typescript-eslint/ban-ts-comment': 'off',
       '@typescript-eslint/no-unsafe-function-type': 'off',
+      'import-x/no-duplicates': 'error',
     },
-    ignores: ['node_modules/*', 'packages/*/dist/*', 'lib/*', '*.html'],
-  }
+    settings: {
+      'import-x/resolver-next': [createTypeScriptImportResolver(), createNodeResolver()],
+    },
+  },
+  globalIgnores(['**/node_modules/**', '**/packages/*/dist/**', '**/lib/**', '**/*.html'])
 )