+++ packages/basic-modules/src/index.ts
@@ -8,6 +8,8 @@ import './assets/index.less'
 // 配置多语言
 import './locale/index'
 
+export * from './modules/types'
+
 import wangEditorParagraphModule from './modules/paragraph'
 import wangEditorTextStyleModule from './modules/text-style'
 import wangEditorHeaderModule from './modules/header'