+++ packages/code-highlight/src/index.ts
@@ -8,6 +8,8 @@ import './assets/index.less'
 // 配置多语言
 import './locale/index'
 
+export * from './custom-types'
+
 import wangEditorCodeHighlightModule from './module/index'
 import wangEditorCodeHighLightDecorate from './decorate'