+++ packages/list-module/src/index.ts
@@ -8,6 +8,8 @@ import './assets/index.less'
 // 配置多语言
 import './locale/index'
 
+export * from './module/custom-types'
+
 // 导出 module
 import wangEditorListModule from './module/index'
 export { wangEditorListModule }