+++ packages/video-module/src/index.ts
@@ -8,5 +8,7 @@ import './assets/index.less'
 // 配置多语言
 import './locale/index'
 
+export * from './module/custom-types'
+
 import wangEditorVideoModule from './module/index'
 export { wangEditorVideoModule }